% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bama.data}
\alias{bama.data}
\title{Synthetic example data for bama}
\format{
A data.frame with 1000 observations on 102 variables:
\describe{
\item{y}{Numeric response variable.}
\item{a}{Numeric exposure variable.}
\item{m[1-100]}{Numeric mediator variables}
}
}
\usage{
bama.data
}
\description{
Synthetic example data for bama
}
\keyword{datasets}
