% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bama-package.R
\docType{package}
\name{bamalog-package}
\alias{bamalog}
\alias{bamalog-package}
\title{bamalog: Batch Behavior Identification for Event Logs}
\description{
Algorithm to identify batching behavior in an event log (both at the task-resource and at the subprocess level)
}
\author{
\strong{Maintainer}: Niels Martin \email{niels.martin@uhasselt.be}

Authors:
\itemize{
  \item Felix Mannhardt \email{felix.mannhardt@sintef.no} [contributor]
}

}
\keyword{internal}
