% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bama_core.R
\name{enumerate_subsequences}
\alias{enumerate_subsequences}
\title{Identify all sequence by enumeration}
\usage{
enumerate_subsequences(task_log, min_trace_freq = 0)
}
\arguments{
\item{task_log}{The task log containing task or activity instances.}

\item{min_trace_freq}{This specifies the number of times that a trace needs to occur in the event log before it should be considered.}
}
\description{
Function which enumerates all subsequences observed within a trace
}
