% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bama_core.R
\name{identify_frequent_sequences}
\alias{identify_frequent_sequences}
\title{Identify frequent sequences}
\usage{
identify_frequent_sequences(task_log, min_support)
}
\arguments{
\item{task_log}{The task log containing task or activity instances.}

\item{min_support}{The minimal support level that needs to be obtained in order to consider a sequence as frequent.}
}
\description{
Function which identifies frequent sequences in traces using the SPADE algoritm (Sequential PAttern Discovery using Equivalence classes)
}
