% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.bama.R
\name{print.fdr.bama}
\alias{print.fdr.bama}
\title{Printing bama objects}
\usage{
\method{print}{fdr.bama}(x, ...)
}
\arguments{
\item{x}{An object of class 'bama'.}

\item{...}{Additional arguments to pass to print.data.frame or summary.bama}
}
\description{
Print a bama object.
}
