% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbm.R
\name{print.hdbm}
\alias{print.hdbm}
\title{Printing hdbm objects}
\usage{
\method{print}{hdbm}(x, ...)
}
\arguments{
\item{x}{An object of class 'hdbm'.}

\item{...}{Additional arguments to pass to print.data.frame or summary.hdbm}
}
\description{
Print a hdbm object.
}
