% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bama_core.R
\name{seq_tolerated_gap_list_generator}
\alias{seq_tolerated_gap_list_generator}
\title{Generate Sequential Tolerated Gap List}
\usage{
seq_tolerated_gap_list_generator(task_log, seq_tolerated_gap_value)
}
\arguments{
\item{task_log}{The task log containing task or activity instances.}

\item{seq_tolerated_gap_value}{This specifies the gap in seconds.}
}
\description{
Function that creates a seq_tolerated_gap_list when a single numeric value (in seconds) is given.
}
