% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbm.R
\name{summary.hdbm}
\alias{summary.hdbm}
\title{Summarise objects of type "hdbm"}
\usage{
\method{summary}{hdbm}(object, rank = F, ci = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{An object of class "hdbm".}

\item{rank}{Whether or not to rank the output by posterior inclusion
probability. Default is TRUE.}

\item{ci}{The credible interval to calculate. \code{ci} should be a length 2
numeric vector specifiying the upper and lower bounds of the CI. By
default, \code{ci = c(0.025, .975)}.}

\item{...}{Additonal optional arguments to \code{summary}}
}
\value{
A data.frame with 4 elements. The beta.m estimates, the estimates'
\emph{credible} interval (which by default is 95%), and their posterior inclusion
probability (pip).
}
\description{
Summarise objects of type "hdbm"
}
