% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bama_api.R
\name{detect_batching}
\alias{detect_batching}
\title{Detect Batch Behavior in an Event Log}
\usage{
detect_batching(
  log,
  subsequence_type = c("mine", "enum"),
  subsequence_min_support = 0.01,
  seq_tolerated_gap = 180,
  within_case_seq_tolerated_gap = seq_tolerated_gap,
  between_cases_seq_tolerated_gap = seq_tolerated_gap,
  arrival_event_gap = NULL
)
}
\arguments{
\item{log}{The `bupaR` event log object that should be animated}
}
\description{
TODO
}
